CAS-173: Port changes from CAS-157 to the PAC4J project
Created a specialized version of SAML2 Client Configuration that holds
data loaded from a relational database. It uses a DAO to lazily load its
properties on init().

Created a specialized version of SAML2 Clients that do not use static
configuration of clients but load them from a database.

The code now depends on PAC4J 1.9.5-SNAPSHOT because some modifications
in PAC4J are needed.